+++ Kernel/Bus/PCI/Initializer.cpp
@@ -4,12 +4,12 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include <Kernel/ACPI/Parser.h>
 #include <Kernel/Bus/PCI/API.h>
 #include <Kernel/Bus/PCI/Access.h>
 #include <Kernel/Bus/PCI/Initializer.h>
 #include <Kernel/Bus/PCI/SysFSPCI.h>
 #include <Kernel/CommandLine.h>
+#include <Kernel/Firmware/ACPI/Parser.h>
 #include <Kernel/IO.h>
 #include <Kernel/Panic.h>
 #include <Kernel/Sections.h>
