+++ Kernel/Devices/MemoryDevice.cpp
@@ -6,8 +6,8 @@
 
 #include <AK/Memory.h>
 #include <AK/StdLibExtras.h>
-#include <Kernel/Arch/PC/BIOS.h>
 #include <Kernel/Devices/MemoryDevice.h>
+#include <Kernel/Firmware/BIOS.h>
 #include <Kernel/Memory/AnonymousVMObject.h>
 #include <Kernel/Sections.h>
