+++ Kernel/Firmware/ACPI/Parser.cpp
@@ -7,11 +7,11 @@
 
 #include <AK/Format.h>
 #include <AK/StringView.h>
-#include <Kernel/ACPI/Parser.h>
-#include <Kernel/Arch/PC/BIOS.h>
 #include <Kernel/Arch/x86/InterruptDisabler.h>
 #include <Kernel/Bus/PCI/API.h>
 #include <Kernel/Debug.h>
+#include <Kernel/Firmware/ACPI/Parser.h>
+#include <Kernel/Firmware/BIOS.h>
 #include <Kernel/IO.h>
 #include <Kernel/Memory/TypedMapping.h>
 #include <Kernel/Sections.h>
