+++ Kernel/Firmware/ACPI/Parser.h
@@ -7,10 +7,10 @@
 #pragma once
 
 #include <AK/Types.h>
-#include <Kernel/ACPI/Definitions.h>
-#include <Kernel/ACPI/Initialize.h>
 #include <Kernel/CommandLine.h>
 #include <Kernel/FileSystem/SysFSComponent.h>
+#include <Kernel/Firmware/ACPI/Definitions.h>
+#include <Kernel/Firmware/ACPI/Initialize.h>
 #include <Kernel/Interrupts/IRQHandler.h>
 #include <Kernel/Memory/Region.h>
 #include <Kernel/PhysicalAddress.h>
