+++ Kernel/Firmware/BIOS.cpp
@@ -6,8 +6,8 @@
  */
 
 #include <AK/StringView.h>
-#include <Kernel/Arch/PC/BIOS.h>
 #include <Kernel/FileSystem/OpenFileDescription.h>
+#include <Kernel/Firmware/BIOS.h>
 #include <Kernel/KBufferBuilder.h>
 #include <Kernel/Memory/MemoryManager.h>
 #include <Kernel/Memory/TypedMapping.h>
