+++ Kernel/Syscalls/shutdown.cpp
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include <Kernel/ACPI/Parser.h>
 #include <Kernel/FileSystem/FileSystem.h>
+#include <Kernel/Firmware/ACPI/Parser.h>
 #include <Kernel/IO.h>
 #include <Kernel/Process.h>
 #include <Kernel/TTY/ConsoleManagement.h>
