+++ Kernel/Time/HPET.cpp
@@ -5,8 +5,8 @@
  */
 
 #include <AK/StringView.h>
-#include <Kernel/ACPI/Parser.h>
 #include <Kernel/Debug.h>
+#include <Kernel/Firmware/ACPI/Parser.h>
 #include <Kernel/Interrupts/InterruptManagement.h>
 #include <Kernel/Memory/MemoryManager.h>
 #include <Kernel/Memory/TypedMapping.h>
