+++ Kernel/Time/TimeManagement.cpp
@@ -7,9 +7,9 @@
 #include <AK/Singleton.h>
 #include <AK/StdLibExtras.h>
 #include <AK/Time.h>
-#include <Kernel/ACPI/Parser.h>
 #include <Kernel/Arch/x86/InterruptDisabler.h>
 #include <Kernel/CommandLine.h>
+#include <Kernel/Firmware/ACPI/Parser.h>
 #include <Kernel/Interrupts/APIC.h>
 #include <Kernel/PerformanceManager.h>
 #include <Kernel/Scheduler.h>
